work on refactoring and fixing clusterOI
